requirements.txt
openai-agents>=0.0.3
pydantic>=2.0.0
flask>=2.3.0
flask-cors>=4.0.0
python-dotenv>=1.0.0
